pyproject.toml
[project]
name = "CPA-Codex-Manager"
version = "1.0.0"
description = "CPA-Codex-Manager"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.0",
    "curl-cffi>=0.14.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pywebview>=5.1",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "psycopg[binary]>=3.1.18",
    "websockets>=16.0",
    "path>=17.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]
payment = [
    "playwright>=1.40.0",
]

[project.scripts]
CPA-Codex-Manager = "webui:main"
CPA-Codex-Manager-Desktop = "desktop:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]
